+++ src/state.ml
@@ -220,7 +220,7 @@ let init window_w window_h window_scale : state =
         settings.music_volume;
       Audio.load_music "daybreak" [ AC_REPAIR_ANNEX ] settings.music_volume;
       Audio.load_music "ending" [ COMPUTER_WING ] settings.music_volume;
-      Audio.load_music "greendale" [ LIBRARY; MEOW_MEOW_BEENZ ] settings.music_volume;
+      Audio.load_music "greendale" [ LIBRARY; MEOW_MEOW_BEENZ; VENTWAYS ] settings.music_volume;
       Audio.load_music "kiss-from-a-rose" ~intro:22.255 ~loop:73.545 [ CITY_OF_CHAIRS; OUTLANDS ]
         settings.music_volume;
       Audio.load_music "mash-theme" ~loop:75.451 [ BASEMENT ] settings.music_volume;